app/requirements.txt
streamlit>=1.30
google-genai>=1.0
python-dotenv
requests
rank_bm25
kiwipiepy
sentence-transformers
